[Update] Favorite create・WantItem createにログイン制限を追加

+++ app/controllers/public/favorites_controller.rb
@@ -8,16 +8,22 @@ class Public::FavoritesController < ApplicationController
   end
 
   def create
-    @favorite = Favorite.new(favorite_params)
-    @favorite.end_user_id = current_end_user.id
-    favorited_item = current_end_user.favorites.find_by(code: @favorite.code)
-    if favorited_item != nil
-      favorited_item.destroy
-      @favorite.save
-      redirect_to end_user_favorites_path(screen_name: current_end_user.screen_name)
+    if end_user_signed_in?
+      @favorite = Favorite.new(favorite_params)
+      @favorite.end_user_id = current_end_user.id
+      favorited_item = current_end_user.favorites.find_by(code: @favorite.code)
+      if favorited_item != nil
+        favorited_item.destroy
+        @favorite.save
+        redirect_to end_user_favorites_path(screen_name: current_end_user.screen_name)
+      else
+        @favorite.save
+        redirect_to end_user_favorites_path(screen_name: current_end_user.screen_name)
+      end
     else
-      @favorite.save
-      redirect_to end_user_favorites_path(screen_name: current_end_user.screen_name)
+      flash[:danger] = "ログインしてください！"
+      redirect_to item_path(params[:favorite][:code], name: params[:favorite][:name], code: params[:favorite][:code],
+            genre: params[:favorite][:genre], price: params[:favorite][:price], image: params[:favorite][:image], url: params[:favorite][:url])
     end
   end
